build: restore test projects to solution and fix compile
Phase 1 excluded test projects from the sln (src compile gate only);
Phase 2a requires the full solution green.

- add UnitTests, Api.UnitTests, Api.IntegrationTests,
  STS.Identity.IntegrationTests to Toralux.Open.IdentityServer.Admin.sln
  (UI.Client.IntegrationTests is an esproj and STS.Identity.ClientTests
  is npm-based; both build under their own tooling)
- remove tests orphaned by Phase 1 feature strip, mirroring server-side
  deletions: IdentityProvider/Key mappers, mocks, and service tests in
  UnitTests and Api.UnitTests (11 files)
- ClientMock: cast DPoPValidationMode enum to int — Open.IdentityServer
  entity stores int, fork-local enum is mapper-only

dotnet build full sln: 0 errors, 0 warnings

+++ tests/Toralux.Open.IdentityServer.Admin.UnitTests/Mocks/ClientMock.cs
@@ -69,7 +69,7 @@ namespace Toralux.Open.IdentityServer.Admin.UnitTests.Mocks
                 .RuleFor(o => o.DPoPClockSkew, f => f.Date.Timespan())
                 .RuleFor(o => o.RequireDPoP, f => f.Random.Bool())
                 .RuleFor(o => o.InitiateLoginUri, f => f.Internet.Url())
-                .RuleFor(o => o.DPoPValidationMode, DPoPTokenExpirationValidationMode.Custom)
+                .RuleFor(o => o.DPoPValidationMode, (int)DPoPTokenExpirationValidationMode.Custom)
                 .RuleFor(o => o.PushedAuthorizationLifetime, f => f.Random.Int())
                 .RuleFor(o => o.RequirePushedAuthorization, f => f.Random.Bool())
                 .RuleFor(o => o.AllowRememberConsent, f => f.Random.Bool())
